fix double to int

+++ ars408_ros/include/ars408_ros/ARS408_CAN.h
@@ -37,24 +37,24 @@ namespace ARS408
     /* 0x201 */
     struct RadarState
     {
-        double NVMReadStatus;
-        double NVMwriteStatus;
-        double MaxDistanceCfg;
-        double Persistent_Error;
-        double Interference;
-        double Temperature_Error;
-        double Temporary_Error;
-        double Voltage_Error;
-        double SensorID;
-        double SortIndex;
-        double RadarPowerCfg;
-        double CtrlRelayCfg;
-        double OutputTypeCfg;
-        double SendQualityCfg;
-        double SendExtInfoCfg;
-        double MotionRxState;
-        double RCS_Threshold;
-        double InvalidClusters;
+        int NVMReadStatus;
+        int NVMwriteStatus;
+        int MaxDistanceCfg;
+        int Persistent_Error;
+        int Interference;
+        int Temperature_Error;
+        int Temporary_Error;
+        int Voltage_Error;
+        int SensorID;
+        int SortIndex;
+        int RadarPowerCfg;
+        int CtrlRelayCfg;
+        int OutputTypeCfg;
+        int SendQualityCfg;
+        int SendExtInfoCfg;
+        int MotionRxState;
+        int RCS_Threshold;
+        int InvalidClusters;
     };
 
     /* 0x203 */
@@ -65,31 +65,30 @@ namespace ARS408
 
     /* 0x204 */
     struct FilterCfg{
-        double Type;
-        double Index;
-        double Active;
+        int Type;
+        int Index;
+        int Active;
         double Min_Distance;
         double Max_Distance;
-
     };
     
     /* 0x408 */
     struct CollDetState{
-        double Activation;
-        double NofRegions;
+        int Activation;
+        int NofRegions;
         double MinDetectTime;
-        double MeasCounter;
+        int MeasCounter;
     };
 
     /* 0x402 */
     struct CollDetRegion{
-        double WarningLevel;
-        double RegionID;
+        int WarningLevel;
+        int RegionID;
         double Point1X;
         double Point1Y;
         double Point2X;
         double Point2Y;
-        double NofObjects;
+        int NofObjects;
         
     };
     
@@ -111,7 +110,7 @@ namespace ARS408
         double DistLat;
         double VrelLong;
         double VrelLat;
-        double DynProp;
+        int DynProp;
         double RCS;
     };
 
@@ -131,22 +130,22 @@ namespace ARS408
         double DistLat;
         double VrelLong;
         double VrelLat;
-        double DynProp;
+        int DynProp;
         double RCS;
     
         /* 0x60C */
         struct Object_quality
         {
             int id;
-            double DistLong_rms;
-            double DistLat_rms;
-            double VrelLong_rms;
-            double VrelLat_rms;
-            double ArelLong_rms;
-            double ArelLat_rms;
-            double Orientation_rms;
-            double ProbOfExist;
-            double MeasState;
+            int DistLong_rms;
+            int DistLat_rms;
+            int VrelLong_rms;
+            int VrelLat_rms;
+            int ArelLong_rms;
+            int ArelLat_rms;
+            int Orientation_rms;
+            int ProbOfExist;
+            int MeasState;
 
         } object_quality;
 
@@ -156,7 +155,7 @@ namespace ARS408
             int id;
             double ArelLong;
             double ArellLat;
-            double Class;
+            int Class;
             double OrientationAngle;
             double Length;
             double Width;
@@ -167,8 +166,7 @@ namespace ARS408
         struct Object_collision
         {
             int id;
-            double CollDetRegionBitfield;
-                        
+            int CollDetRegionBitfield;
         } object_collision;
     };
 }
